projects/DMML#04/src/requirements.txt
sentence-transformers>=2.2.2
pandas>=1.3.5
numpy>=1.21.6
scikit-learn>=1.0.2
umap-learn>=0.5.3
matplotlib>=3.5.3
torch>=1.12.1
tqdm>=4.64.0